delta_joint: verify on real dataset — correct 25x claim to measured 6.5x net
Ran the pipeline on the real lerobot_dataset (2 eps, 5401 frames, 30 Hz) and
printed actual samples. Everything held except two of my own numbers:

- roundtrip is exact to 3e-8 deg (fp32 (a-s)+s rounding), not bitwise — the
  smoke set happened to be bitwise-clean, real magnitudes (~145 deg) are not
- the '5x times 5x = 25x' gain claim double-counted: stride widens the chunk to
  8.2 s, far entries move more, and the normaliser std grows with them
  (3.0 -> 8.2 deg). Measured net: signal/std 0.005 -> 0.032 = 6.5x. README and
  notebook corrected.
- stats-vs-getitem verified exact on all 270k real chunk-entries (4e-7 in
  float64); test tightened to float64 accumulation so it stays valid at scale

+++ delta_joint/test_delta_joint.py
@@ -78,7 +78,9 @@ def test_stats_match_getitem():
     included — a mismatch silently mis-normalises every target."""
     for stride in (1, STRIDE):
         ds = _ds(action_stride=stride)
-        seen = np.concatenate([ds[i]["action"].numpy() for i in range(len(ds))])
+        # float64: np.std in float32 accumulates ~2e-3 error over the real
+        # dataset's 270k chunk-entries (measured) — get_stats itself is exact
+        seen = np.concatenate([ds[i]["action"].numpy() for i in range(len(ds))]).astype(np.float64)
         st = ds.get_stats()
         np.testing.assert_allclose(seen.mean(0), st["action_mean"], atol=1e-3)
         np.testing.assert_allclose(seen.std(0).clip(1e-6), st["action_std"], atol=1e-3)
